pyproject.toml
[project]
name = "mcpdocs"
version = "0.1.0"
description = "Sample MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.3",
    "python-dotenv>=1.0.0",
    "requests>=2.32.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sample_mcp_server"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
# Relaxed type checking for compatibility with MCP library decorators
# Enable strict mode once MCP library has complete type stubs
disallow_untyped_defs = false
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
